Pre-generate trig tables. Remove wall/roof rendering.
git-svn-id: svn+ssh://jay/var/svn/wolf/trunk@8 32837ae5-38f0-4cfd-8401-3ff76d8497c4

+++ wl_def.h
@@ -692,7 +692,7 @@ extern	myint		mouseadjustment;
 
 /* math tables */
 extern myint pixelangle[MAXVIEWWIDTH];
-extern long finetangent[FINEANGLES/4];
+extern int32_t finetangent[FINEANGLES/4];
 extern fixed sintable[], *costable;
 
 extern char configname[13];
